+++ Ferramax/package.json
@@ -18,15 +18,13 @@
     "knex": "^3.1.0",
     "pg": "^8.11.5",
     "remult": "^0.26.10",
-    "vue": "^3.4.21",
-    "vue-router": "^4.0.13"
+    "vue": "^3.4.21"
   },
   "devDependencies": {
     "@tsconfig/node20": "^20.1.4",
     "@types/cookie-session": "^2.0.49",
     "@types/express": "^4.17.21",
-    "@types/node": "^20.12.12",
-    "@types/vue-router": "^2.0.0",
+    "@types/node": "^20.12.10",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vue/tsconfig": "^0.5.1",
     "npm-run-all2": "^6.1.2",
