+++ Ferramax/src/main.ts
@@ -1,7 +1,8 @@
 // src/main.ts
 
-import { createApp } from 'vue';
-import App from './App.vue';
-import router from './router/index';
+import { createApp } from "vue"
+import App from "./App.vue"
 
-createApp(App).use(router).mount('#app');
+import "./assets/main.css"
+
+createApp(App).mount("#app")
