--- Ferramax/src/views/About.vue
@@ -1,17 +1,0 @@
-<template>
-    <div class="about">
-      <h1>About Page</h1>
-      <p>Esta es la página sobre nosotros.</p>
-    </div>
-  </template>
-  
-  <script lang="ts">
-  import { defineComponent } from 'vue';
-  
-  export default defineComponent({
-    name: 'About'
-  });
-  </script>
-  
-
-
