--- Ferramax/src/views/Home.vue
@@ -1,15 +1,0 @@
-<template>
-    <div class="home">
-      <h1>Home Page</h1>
-      <p>Bienvenido a la página principal.</p>
-    </div>
-  </template>
-  
-  <script lang="ts">
-  import { defineComponent } from 'vue';
-  
-  export default defineComponent({
-    name: 'Home'
-  });
-  </script>
-
